requirements.txt
discord.py==2.5.2
Flask>=3.1.1
python-dotenv
PyNaCl
aiohttp>=3.9.1
Pillow
beautifulsoup4
bs4
openai>=1.12.0
TikTokLive>=5.0.6